delete plans fixed

+++ src/Traits/ManagesPlan.php
@@ -140,21 +140,7 @@ trait ManagesPlan
             $options['name'] = $name;
         }
 
-        if (! array_key_exists('description', $options) && $description = $this->description) {
-            $options['description'] = $description;
-        }
-
-        if (! array_key_exists('amount', $options) && $amount = $this->amount) {
-            $options['amount'] = $amount;
-        }
-
-        if (! array_key_exists('interval', $options) && $interval = $this->interval) {
-            $options['interval'] = $interval;
-        }
-
-        if (! array_key_exists('currency', $options) && $currency = $this->currency) {
-            $options['currency'] = $currency;
-        }
+        if(null == $this->plan_code) return;
 
         return PaystackPlan::delete(
             $this->plan_code, $options, $this->paystackOptions()
